move tests into its own files
